immer noch manage tb

+++ manage_table.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 	source /home/uwe/my_scripts/my_functions.sh
 
+/home/uwe/my_scripts/dbms.sh --func ctrl_manage_tb
+exit
+
 function create_tb () {
+
 	db="$1";tb="$2" 
-##### load table info for tb
+##### create tmp table to store table-info
 	crtb="edit_$tb"
 	echo "drop table if exists $crtb;" > $tmpf2  
 	echo "create table   $crtb ( " \
@@ -16,12 +20,12 @@ function create_tb () {
 	     "ref_field		 text,	 ref_table 	    text,	on_delete  	 text,	on_update   	 text);" >> $tmpf2  
 	echo "insert into $crtb (pos,field,type,nullable,default_value,primarykey) values" >> $tmpf2
 ###	table info
-	echo "pragma table_info($tb);" | sqlite3 "$db" |  tr -d '\r' |  tr '[:upper:]' '[:lower:]' |
+	func_sql_execute "$db" "pragma table_info($tb)" |  tr '[:upper:]' '[:lower:]' |
 	while read -r line;do
 	    IFS=",";fields=( $line );unset IFS;nline="";del=""
 	    for ((ia=0;ia<${#fields[@]};ia++)) ;do
 			arr=$(echo ${fields[$ia]} | tr -d '"' | tr -d "'")
-			if [ "$arr" = "0" ];then arr=""  ;fi
+			if [ "$arr" = "0" ] && [ "$ia" != "0" ] ;then arr=""  ;fi
 			case "$ia" in
 				3)		if [ "$arr"  = "1" ];then  arr="not null" ;fi;;
 				5)	    if [ "$arr"  = "1" ];then  arr="primary key";fi;;
@@ -33,11 +37,12 @@ function create_tb () {
 		delim=","
 	done 
 	echo ";" >> $tmpf2
+	set -x
 ###	index info
-	echo "pragma index_list($tb);" | sqlite3 "$db" |  tr -d '\r' | tr '[:upper:]' '[:lower:]' |
+	func_sql_execute "$db" "pragma index_list($tb)" |  tr '[:upper:]' '[:lower:]' |
     while read line; do
 		IFS=",";arr=($line);printf "${arr[1]},${arr[2]},${arr[3]},"
-		echo "pragma index_info(${arr[1]});" | sqlite3 "$db" |  tr -d '\r' |  tr '[:upper:]' '[:lower:]'
+		func_sql_execute "$db" "pragma index_info(${arr[1]})" |  tr '[:upper:]' '[:lower:]'  
 	done |	
     while read line; do
 		IFS=",";arr=($line);unset IFS;del=","
@@ -48,13 +53,15 @@ function create_tb () {
 	done 
 	echo "update $crtb set auto_increment = \"autoincrement\" where primarykey = 'primary key' and type = 'integer';" >> $tmpf2
 ###	foreign key info
-	echo "pragma foreign_key_list($tb);" | sqlite3 "$db" |  tr -d '\r' |  tr -d '"' | tr '[:upper:]' '[:lower:]' |
+	func_sql_execute "$db" "pragma foreign_key_list($tb)" |  tr '[:upper:]' '[:lower:]' |
     while read line; do
 		IFS=",";arr=($line) 
 		echo "update $crtb set ref_table = \"${arr[2]}\", ref_field = \"${arr[4]}\"," \
 			 "on_update = \"${arr[5]}\", on_delete = \"${arr[6]}\" where field = \"${arr[3]}\";" >> $tmpf2
 	done  
-	echo ".read $tmpf2" | sqlite3 "/home/uwe/my_databases/parm.sqlite"
+	setmsg -i pause
+    func_sql_execute "/home/uwe/my_databases/parm.sqlite" ".read $tmpf2"
+    if [ "$?" -gt "0" ];then return 1;fi
 ##### user action
 	dbms.sh "/home/uwe/my_databases/parm.sqlite" $crtb "--notable" 1> /dev/null
 ##### create file for .read
@@ -62,8 +69,9 @@ function create_tb () {
 	echo "    create table $tb ("
 	export del="   "
 	[ -f "$tmpf2" ] && rm $tmpf2
-	echo "select field,type,primarykey,auto_increment,nullable,isunique,default_value,ixname,ref_table,ref_field,on_delete,on_update,pos from $crtb;" | 
-	sqlite3 	"/home/uwe/my_databases/parm.sqlite" | 	tr -d '\r' |  tr -d '"' |
+	stmt="select field,type,primarykey,auto_increment,nullable,isunique,default_value,
+		 ixname,ref_table,ref_field,on_delete,on_update,pos from $crtb;"  
+	func_sql_execute "/home/uwe/my_databases/parm.sqlite" $stmt  |  tr -d '"' |
 	while read -r line;do
 		IFS=",";fields=( $line );unset IFS;nline="$del";if [ "$nline" = "" ];then nline="       "  ;fi
 	    for ((ia=0;ia<${#fields[@]};ia++)) ;do
@@ -74,7 +82,6 @@ function create_tb () {
 				*)  
 			esac
 			case "$ia" in
-#				5)	   if [ "${fields[7]}"  != "" ];then  arr="";fi;;
 				6)	   if [ "$arr"  != "" ];then  arr="default \"$arr\"";fi;;
 				7)	   if [ "$arr"  != "" ];then
 						   if [ "${fields[5]}"  != "" ];then  
@@ -135,42 +142,50 @@ function create_tb () {
 }
 	db="$1";tb=$2;func="$3";true=0;false=1
 #	db="/home/uwe/my_databases/test.sqlite";tb="mytable_neu";func="edit"
-	db="/home/uwe/my_databases/music.sqlite";tb="track";func="edit"
-	if [ "$(echo $func | grep 'drop')"	 != "" ]; 		then drop=$true;	else drop=$false;	fi 
-	if [ "$(echo $func | grep 'create')"	 != "" ]; 	then create=$true;	else create=$false;	fi 
-	if [ "$(echo $func | grep 'edit')"	 != "" ]; 		then edit=$true;create=$true;	else edit=$false;	fi 
-	if [ "$(echo $func | grep 'import')"	 != "" ]; 	then import=$true;	else import=$false;	fi 
+	db="/home/uwe/my_databases/music.sqlite";tb="track";nfunc="edit"
+	if [ "$func" = "" ]; then func=$(zenity --list --column action "drop" "schema" "modify" "new_table" "import");fi
+	if [ "$(echo $func | grep 'drop')"	 	!= "" ]; 	then drop=$true;				else drop=$false;	fi 
+	if [ "$(echo $func | grep 'schema')" 	!= "" ]; 	then create=$true;				else create=$false;	fi 
+	if [ "$(echo $func | grep 'modify')"	!= "" ]; 	then edit=$true;create=$true;	else edit=$false;	fi 
+	if [ "$(echo $func | grep 'new_table')" != "" ]; 	then tb=$(zenity --text "enter table name" --entry) ;fi 
+	if [ "$(echo $func | grep 'new_table')" != "" ]; 	then edit=$true;create=$true;newtable=$true;	else edit=$false;newtable=$false;fi 
+	if [ "$(echo $func | grep 'import')" != "" ]; 		then import=$true;	else import=$false;	fi 
 	if [ "$db" = "" ];then db=$(dbms.sh --func get_fileselect parm_value searchpath database --save);fi
 	if [ "$db" = "" ];then setmsg -n "abort..no db selected"; exit ;fi
-	if [ -f "$db" ] && [ "$tb" = "" ]; then tb=$(zenity --text "neu ueberschreiben " --list --editable --column tabelle $(echo ".tables" | sqlite3 $db) "neu");fi 
-	if [ "$tb" = "" ];then tb=$(zenity --text "neue Tabelle" --entry) ;fi
+	if [ -f "$db" ] && [ "$tb" = "" ]; then tb=$(zenity --text "neu ueberschreiben " --list --editable --column tabelle $(echo ".tables" | sqlite3 $db) "new");fi 
+	if [ "$tb" = "" ] || [ "$tb" = "new" ];then tb=$(zenity --text "neue Tabelle" --entry) ;fi
 	if [ "$tb" = "" ];then setmsg -n "abort..no tb selected"; exit ;fi
-	if [ "$func" = "" ]; then func=$(zenity --list --column action "drop" "create" "import");fi
 	tmpf="/tmp/create_tb.txt"
 	tmpf2="/tmp/create_tb2.txt"
 	import="/home/uwe/.dbms/import/my_table.csv"
-	
 	if [ "$create" = "$true" ]; then
 		found=$(echo ".tables $tb" | sqlite3 $db)
 		if [ "$found" = "" ]; then
-			echo "create table $tb (${tb}_id  integer primary key autoincrement not null unique,${tb}_name	text);" |	sqlite3 "$db"
+			func_sql_execute "$db" "create table $tb (${tb}_id  integer primary key autoincrement not null unique,${tb}_name	text);"
 			edit=$true
+		else
+			newtable=$false
 		fi
 		func_tb_meta_info $db $tb;TINSERT=$TSELECT
-		echo "	drop table   if     exists ${tb}_copy;" 		>  $tmpf
-		echo "	alter table $tb rename to ${tb}_copy;"	>> $tmpf
+		if [ "$newtable" != "$true" ];then 
+			echo "	drop table   if     exists ${tb}_copy;" 		>  $tmpf
+			echo "	alter table $tb rename to ${tb}_copy;"	>> $tmpf
+		fi
 		if [ "$drop" = "$true" ]; then echo "	drop table   if     exists $tb;" >> $tmpf;fi
 		if [ "$edit" = "$true" ]; then		
 			create_tb "$db" "$tb" 								>> $tmpf
 		else
 			echo  ".schema $tb" | sqlite3 $db 					>> $tmpf 
 		fi
-		echo "	insert into $tb  ($TSELECT) " 					>> $tmpf
-		echo "	select            $TSELECT " 					>> $tmpf
-		echo "	from ${tb}_copy;" 								>> $tmpf
+		if [ "$newtable" != "$true" ];then 
+			echo "	insert into $tb  ($TSELECT) " 				>> $tmpf
+			echo "	select            $TSELECT " 				>> $tmpf
+			echo "	from ${tb}_copy;" 							>> $tmpf
+		fi
 		xdg-open $tmpf
 		setmsg -q "$tmpf ausfuehren?" 
 		if [ "$?" = "1" ];then 
+			if [ "$newtable" = "$true" ];then func_sql_execute "$db" "drop table if exists $tb;";exit ; fi
 			setmsg -n Abbruch;exit
 		else	
 			echo ".read $tmpf" | sqlite3 $db
